Unpin ruby version in Gemfile
The .ruby-version file specifies 3.3.1, but pulling this into the
Gemfile pins this version exactly, which means we don't benefit from
security fixes in later patch versions.

The bundler documentation suggests that patchlevel is optional for this
exact reason - https://bundler.io/guides/gemfile_ruby.html - and
supports the twiddle wakka syntax (~>) which here is equivalent to ">=
3.3.1", "< 3.4.0".

Sadly this means we have to go back to duplicating the version
constraint in .ruby-version and the Gemfile, but c'est la vie.